bailam

+++ module4/6.JPA/exercise/blog/src/main/java/nhat/coder/repository/BlogRepository.java
@@ -8,6 +8,6 @@ public interface BlogRepository {
     void save(Blog blog);
     List<Blog> getAll();
     public Blog findById(Long id);
-    public void update(int id, Blog blog);
+    public void update(Long id, Blog blog);
     void remove(Blog blog);
 }
