+++ module4/6.JPA/exercise/blog/src/main/java/nhat/coder/service/BlogService.java
@@ -7,7 +7,8 @@ import java.util.List;
 public interface BlogService {
     void save(Blog blog);
     List<Blog> getAll();
+
     public Blog findById(Long id);
-    public void update(int id, Blog blog);
+    public void update(Long id, Blog blog);
     void remove(Blog blog);
 }
